refactor package of assert.statement
