+++ shardingsphere-sql-parser/shardingsphere-sql-parser-test/src/test/java/org/apache/shardingsphere/sql/parser/integrate/asserts/statement/SQLStatementAssert.java
@@ -21,15 +21,15 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import org.apache.shardingsphere.sql.parser.integrate.asserts.SQLCaseAssertContext;
 import org.apache.shardingsphere.sql.parser.integrate.asserts.segment.parameter.ParameterMarkerAssert;
-import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl.AlterTableStatementAssert;
-import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl.CreateIndexStatementAssert;
-import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl.CreateTableStatementAssert;
-import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl.DeleteStatementAssert;
-import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl.DropIndexStatementAssert;
-import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl.InsertStatementAssert;
-import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl.SelectStatementAssert;
-import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl.TCLStatementAssert;
-import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.impl.UpdateStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.AlterTableStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.CreateIndexStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.CreateTableStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.dml.DeleteStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.ddl.DropIndexStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.dml.InsertStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.dml.SelectStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.tcl.TCLStatementAssert;
+import org.apache.shardingsphere.sql.parser.integrate.asserts.statement.dml.UpdateStatementAssert;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.SQLParserTestCase;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.ddl.AlterTableStatementTestCase;
 import org.apache.shardingsphere.sql.parser.integrate.jaxb.statement.ddl.CreateIndexStatementTestCase;
